% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somobj_doc.R
\name{as_list}
\alias{as_list}
\title{Convert an SOM object to a list}
\usage{
SOMobj$as_list()
}
\value{
A list
}
\description{
This method extracts all fields of an SOM object and places their stored values 
into the fields of an R list, with list field names matching the SOM object field names.
}
