% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{build_CADJ}
\alias{build_CADJ}
\title{Build a CADJ matrix}
\usage{
build_CADJ(BMU, nW, parallel = TRUE)
}
\arguments{
\item{BMU}{matrix (nrow = nrow(X), ncol = nBMU) of indices (1-based) of the BMUs of each data vector in X.}

\item{nW}{the number of SOM prototypes (CADJ matrices have nrow = ncol = nW)}

\item{parallel}{optional, whether to compute in parallel. Default = TRUE}
}
\value{
a CADJ matrix
}
\description{
Build a CADJ matrix
}
\details{
The input BMU matrix should be the result of calling \code{find_BMU} with \code{nBMU} >= 2.
}
