% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_tools.R
\name{build_ctab}
\alias{build_ctab}
\title{Build a color table}
\usage{
build_ctab(labels)
}
\arguments{
\item{labels}{a set of data labels to whose unique values colors will be assigned}
}
\description{
Returns a default \code{ctab} given a set of (possibly non-unique) labels. 
The returned object is a data frame with columns \code{label} and \code{color} defining the color mapping.
}
