% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_SOM_fences}
\alias{calc_SOM_fences}
\title{Calculate the fences of the SOM lattice}
\usage{
calc_SOM_fences(nu_xy, W, nu_verts, nu_ADJ, parallel = TRUE)
}
\arguments{
\item{nu_xy}{a matrix whose rows contain the (x,y) coordinates (cols 1 and 2, respectively) of each lattice neuron, ordered 
from the bottom-left of the lattice.}

\item{W}{the prototype matrix (rows ordered the same as nu_xy)}

\item{nu_verts}{a 3-d array whose slices contain the vertices of each lattice tile (slices ordered the same as nu_xy)}

\item{nu_ADJ}{the neuron adjacency matrix}

\item{parallel}{optional, whether to compute in parallel. Default = TRUE}
}
\value{
A data frame identifying the fence and its corresponding value between each pair of neighboring lattice neurons, with columns: 
\itemize{
\item i the index of the 1st neuron comprising the fence 
\item j the index of the 2nd neuron comprising the fence
\item x0 the lattice x-coord of neuron i 
\item y0 the lattice y-coord of neuron i 
\item x1 the lattice x-coord of neuron j 
\item y1 the lattice y-coord of neuron j
\item value the fence value separating neurons i and j 
}
}
\description{
Lattice fences between neighboring neurons on the lattice are the (squared) Euclidean distance between 
the neurons' corresponding prototypes.
}
