% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somobj_doc.R
\name{CONN}
\alias{CONN}
\title{The CONNectivity matrix of SOM prototypes}
\usage{
SOMobj$CONN()
}
\value{
the CONN matrix (nrow = ncol = \code{nW})
}
\description{
The CONN matrix is calculated from the \code{CADJ} matrix computed during \link{recall_SOM}: 
\deqn{CONN = CADJ + t(CADJ)}. 

\code{CADJ} describes the strength of the topological connectivites between SOM prototypes in input space 
(higher CADJ values between prototypes \eqn{w_i} and \eqn{w_j} indicate stronger evidence that \code{w_i} and \code{w_j} belong to the same cluster). 
The (i,j) values of CADJ are defined: 
\deqn{CADJ_{ij} = \#(BMU1(X)=i and BMU2(X)=j)}

By this definition, \code{CADJ} is an asymmetric adjacency matrix. CONN is its symmetric counterpart.
}
\references{
\insertRef{TasdemirMerenyi2009}{SOMDisco}
}
