% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRAS.R
\name{default_LRAS}
\alias{default_LRAS}
\title{Generate a default Learning Rate Annealing Schedule}
\usage{
default_LRAS(nX)
}
\arguments{
\item{nX}{the number of data vectors used for training}
}
\value{
the Learning Rate Annealing Schedule as a data frame with columns
t (time step), 
alpha (prototype update rate), 
beta (win frequencies update rate), 
gamma (bias update rate), 
sigma (geodesic lattice distance to which prototype update is applied)
}
\description{
Generate a default Learning Rate Annealing Schedule
}
