% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distmat}
\alias{distmat}
\title{Compute distances between rows of two matrices}
\usage{
distmat(
  X1,
  X2,
  which_dist = "L22",
  parallel = TRUE,
  bias = NULL,
  X1min = NULL,
  X1max = NULL,
  X2min = NULL,
  X2max = NULL
)
}
\arguments{
\item{X1}{a data matrix, vectors in rows}

\item{X2}{a data matrix, vectors in rows. ncol(X2) must equal ncol(X1).}

\item{which_dist}{optional, the distance to compute. Options are \itemize{\item "L2" for Euclidean, \item "L22" for Squared Euclidean", \item "L1" for L-1 distance, \item "LInf" for L-Inf distance} Default = "L22".}

\item{parallel}{optional, whether to compute in parallel. Default = TRUE}

\item{bias}{optional, a vector of bias to be applied to the distance calculation when computing BMU. 
Default = NULL, meaning no bias is applied. If given, length(bias) must equal nrow(X2) (one bias for every row of X2).}

\item{X1min}{optional, a vector giving the minimum of the range of X1 used in any linearly scaling before distances are computed. 
Default = NULL, meaning no scaling is applied.}

\item{X1max}{optional, a vector giving the maximum of the range of X1 used in any linearly scaling before distances are computed. 
Default = NULL, meaning no scaling is applied.}

\item{X2min}{optional, a vector giving the minimum of the range of X2 used in any linearly scaling before distances are computed. 
Default = NULL, meaning no scaling is applied.}

\item{X2max}{optional, a vector giving the maximum of the range of X2 used in any linearly scaling before distances are computed. 
Default = NULL, meaning no scaling is applied.}
}
\value{
a matrix whose (i,j) entry contains the requested distance between the i-th row of X1 and and j-th row of X2
}
\description{
Compute distances between rows of two matrices
}
\details{
If given, bias affects the distance calculation between the rows of X1 and X2 by: which_dist(X1[i,], X2[j,]) - bias[j]. 

The inputs X1min, X1max, X2min, X2max control any linearly scaling (from the range of X1, to that of X2) that is applied prior to distance calculation. 
If any of the above are given, they all must be given.  Each can be given as either a vector controlling ranges by dimension (length = ncol(X1)), 
or a length 1 vector (in which case the single value will be recycled across dimensions).  

Specifying scaling forces distance computation to be performed in the X2 range. This is an optional feature.
}
