% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{find_BMU}
\alias{find_BMU}
\title{Find BMU of data vectors}
\usage{
find_BMU(
  X,
  W,
  nBMU = 2L,
  parallel = TRUE,
  bias = NULL,
  Xmin = NULL,
  Xmax = NULL,
  Wmin = NULL,
  Wmax = NULL
)
}
\arguments{
\item{X}{a data matrix, vectors in rows}

\item{W}{a prototype matrix, vectors in rows}

\item{nBMU}{optional, the number of BMUs returned. Default = 2.}

\item{parallel}{optional, whether to compute in parallel. Default = TRUE}

\item{bias}{optional, a vector of bias to be applied to the distance calculation when computing BMU. 
Default = NULL, meaning no bias is applied. If given, length(bias) must equal nrow(W) (one bias for every prototype).}

\item{Xmin}{optional, a vector giving the minimum of the range of X used in any linearly scaling before distances are computed. 
Default = NULL, meaning no scaling is applied.}

\item{Xmax}{optional, a vector giving the maximum of the range of X used in any linearly scaling before distances are computed. 
Default = NULL, meaning no scaling is applied.}

\item{Wmin}{optional, a vector giving the minimum of the range of W used in any linearly scaling before distances are computed. 
Default = NULL, meaning no scaling is applied.}

\item{Wmax}{optional, a vector giving the maximum of the range of W used in any linearly scaling before distances are computed. 
Default = NULL, meaning no scaling is applied.}
}
\value{
a list with components 
\itemize{ 
  \item BMU a matrix (nrow = nrow(X), ncol = nBMU) whose (i,j) entry lists the index (the row of W, 1-based) of the j-th BMU of the i-th row of X. 
  \item SQE a vector (length = nrow(X)) whose i-th element gives the Squared Quantization Error of the i-th row of X
}
}
\description{
Find BMU of data vectors
}
\details{
If given, bias affects the distance calculation between the rows of X and W by: dist(X[i,], W[j,]) - bias[j]. 

The inputs Xmin, Xmax, Wmin, Wmax control any linearly scaling that is applied prior to distance calculation. 
If any of the above are given, they all must be given.  Each can be given as either a vector controlling ranges by dimension (length = ncol(W)), 
or a length 1 vector (in which case the single value will be recycled across dimensions).  Specifying the X and W ranges is only useful 
if you are trying to match a BMU selection that was performed during SOM training, which utilized an external / internal network range.
}
