% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{find_RF_label}
\alias{find_RF_label}
\title{Compute the plurality label of each prototype.}
\usage{
find_RF_label(X_label, RF_members, parallel = TRUE)
}
\arguments{
\item{X_label}{a vector (length = nrow(X)) of data labels}

\item{RF_members}{a list containing the (1-based) indices of the data vectors mapped to each prototype, 
as returned by \code{find_RF_members}.}

\item{parallel}{optional, whether to compute in parallel. Default = TRUE}
}
\value{
a vector of prototype labels. If a prototype's Receptive Field is empty the returned label is NA
}
\description{
For labeled data mapped to each prototype's Receptive Field, the plurality label
is returned.
}
