% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{find_RF_members}
\alias{find_RF_members}
\title{Produce a list of data vectors in each prototype's Receptive Field}
\usage{
find_RF_members(BMU, nW, parallel = TRUE)
}
\arguments{
\item{BMU}{matrix (nrow = nrow(X), ncol = nBMU) of indices (1-based) of the BMUs of each data vector in X.}

\item{nW}{the number of SOM prototypes (CADJ matrices have nrow = ncol = nW)}

\item{parallel}{optional, whether to compute in parallel. Default = TRUE}
}
\value{
a list (length = nW), containing the (1-based) indices of the data vectors mapped to each prototype.
}
\description{
Produce a list of data vectors in each prototype's Receptive Field
}
\details{
The input BMU matrix should be the result of calling \code{find_BMU} with \code{nBMU} >= 2.
}
