% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{geodesicdist}
\alias{geodesicdist}
\title{Shortest path distances between vertices of a graph}
\usage{
geodesicdist(ADJ, weighted = FALSE, directed = FALSE)
}
\arguments{
\item{ADJ}{the adjacency matrix of graph vertices. Can be weighted.}

\item{weighted}{optional, whether to consider the edge weights found in ADJ and construct a weighted graph. Default = FALSE.}

\item{directed}{optional, whether to consider the graph as directed. Default = FALSE.}
}
\value{
a shortest-path distance matrix between graph vertices (nrow = ncol = num. vertices)
}
\description{
This is a wrapper for igraph::distances, see its help for more information.
}
