% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somobj_doc.R
\name{get_LRAS}
\alias{get_LRAS}
\title{Get the Learning Rate Annealing Schedule}
\usage{
SOMobj$get_LRAS()
}
\value{
the effective LRAS of the SOM, as a data frame.
}
\description{
A getter function for the LRAS (described in \link{set_LRAS}) stored in the SOM object.
}
