% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somobj_doc.R
\name{initialize_SOM}
\alias{initialize_SOM}
\title{Initialize a SOM object}
\usage{
SOMobj$initialize_SOM(X, som_x, som_y, lattice_type)
}
\arguments{
\item{X}{matrix of training data (data vectors in rows)}

\item{som_x}{width of the SOM (number of neurons)}

\item{som_y}{height of the SOM (number of neurons)}

\item{lattice_type}{either "grid" for rectangular lattices or "hex" for hexagonal lattices}
}
\value{
None, data and lattice related fields are set internally
}
\description{
Sets up a SOM lattice for training, populates default training parameters, and attaches the training data to the specified SOM.
}
\details{
This is a wrapper function to perform all steps necessary to setup a SOM for learning. 
Internally it calls \link{set_lattice}, \link{set_netrng}, \link{set_W_runif}, \link{set_p_equal}, and \link{set_LRAS}, all with their default arguments.  
It also sets internal variables: 
\itemize{
\item \code{som_x}, \code{som_y}, \code{nW}, \code{lattice_type}
\item \code{d}, \code{nX}, \code{X_stats}
\item \code{nBMU}
}
}
