% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{linscale}
\alias{linscale}
\title{Linear scaling of the rows of a data matrix}
\usage{
linscale(X, from_min, from_max, to_min, to_max, parallel = TRUE)
}
\arguments{
\item{X}{a data matrix with vectors in rows}

\item{from_min}{a vector defining the min from which X is scaled, by dimension}

\item{from_max}{a vector defining the max from which X is scaled, by dimension}

\item{to_min}{a vector defining the min to which X is scaled, by dimension}

\item{to_max}{a vector defining the max to which X is scaled, by dimension}

\item{parallel}{boolean, whether to compute in parallel}
}
\value{
a matrix (same dimensions as X) whose rows are scaled to the requested range
}
\description{
Each row is mapped linearly from a given range, to a given range
}
