% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somobj_doc.R
\name{load}
\alias{load}
\title{Load an existing SOM object}
\usage{
SOMobj$load(somfile)
}
\arguments{
\item{somfile}{a string indicating the file path and name of the saved SOM object.}
}
\value{
None, the SOM object is loaded
}
\description{
SOM objects previously written to disk via the \link{save} method can be re-loaded into 
a new R environment with this function. All fields of the internal C++ class will be populated, and 
all methods can be called on the loaded SOM object.
}
\details{
Because the .som file is in .rds format it can, technically, be loaded directly into an R environment 
as a list via \link[base]{readRDS}. This can be useful for spot checking the contents of a saved SOM object, but 
does not allow use of any of its methods (or visualizations).  The \code{load} methods allows for 
proper restoration of a previously saved SOM.
}
