% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somobj_doc.R
\name{load_list}
\alias{load_list}
\title{Populate a SOM object from a list}
\usage{
SOMobj$load_list(SOMList)
}
\arguments{
\item{SOMList}{a SOM object converted to a list, e.g., with \code{as_list}}
}
\value{
None
}
\description{
This method populates all fields of a SOM object from the fields of an R list object. 
The list must have field names which exactly match the SOM field names.
}
