% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somobj_doc.R
\name{map_from_netrng}
\alias{map_from_netrng}
\title{Map a prototype matrix to external network range}
\usage{
SOMobj$map_from_netrng(W)
}
\arguments{
\item{W}{a matrix occupying internal network range to be mapped to external network range. 
\code{W} can be either a single vector, or a matrix with multiple vectors in its rows. 
It must have the same dimension (number of columns) as the SOM (which is stored internally as \code{d}). 

Note that the SOM prototypes are computed and stored in internal network range; to compare them directly 
to the training data they should be mapped to the external range via this function.}
}
\value{
a matrix (same dimension as input) containing the scaled data vectors in its rows
}
\description{
Provides reverse functionality of \link{map_to_netrng}, to linearly scale vectors from internal to external network range.
}
