% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somobj_doc.R
\name{map_to_netrng}
\alias{map_to_netrng}
\title{Map a data matrix to internal network range}
\usage{
SOMobj$map_to_netrng(X)
}
\arguments{
\item{X}{a data matrix (occupying the external network range) to be mapped to the internal network range. 
\code{X} can be either a single vector, or a matrix with multiple data vectors in its rows. 
It must have the same dimension (number of columns) as the SOM (which is stored internally as \code{d}).}
}
\value{
a matrix (same dimension as input) containing the scaled data vectors in its rows
}
\description{
During training, prototype updates and BMU selelction are performed in the internal network range.  
This function allows mapping an arbitrary data matrix to this range to, e.g., replicate BMU selection or 
be able to compare data to prototypes.
}
