% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somobj_doc.R
\name{new}
\alias{new}
\title{Create an empty SOM object}
\usage{
SOM$new()
}
\value{
an empty SOM object templated for initialization (via \link{initialize_SOM}) and training (via \link{train_SOM}).
}
\description{
Create an empty SOM object
}
