% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somobj_doc.R
\name{save}
\alias{save}
\title{Save a SOM object}
\usage{
SOMobj$save(somfile)
}
\arguments{
\item{somfile}{a string indicating the file path and name in which to save the SOM object.
This must end in extension ".som", otherwise an error is returned.}
}
\value{
None, the SOM object is saved to disk
}
\description{
All fields in a SOM object can be saved to disk with this function, which allows them to be re-loaded 
into a new R environment at a later time (for analysis, or possibly extended training).
}
\details{
The SOM object is saved to disk as an R list, with each field occupying a corresponding field of the list.  
The file is saved in .rds format (can check its details with \link[base]{infoRDS}).  

Saved SOMs can be re-loaded with \link{load}
}
