% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somobj_doc.R
\name{set_lattice}
\alias{set_lattice}
\title{Setter function for the SOM lattice.}
\usage{
SOMobj$set_lattice()
}
\value{
None, lattice related fields are set internally
}
\description{
Sets the neuron lattice coordinates and associated quantities needed for SOM training and visualization.
}
\details{
\link{initialize_SOM} must be called prior to calling \code{set_lattice}. 

Internally, the following fields are computed and set: 
\itemize{
\item \code{nu_xy} the (x,y) coordinates of neurons on the lattice
\item \code{nu_ij} the (i=row,j=col) coordinates of neurons on the lattice
\item \code{nW} the number of neurons / prototypes
\item \code{nu_ADJ} the adjacency matrix of neurons on the lattice (identifying immediate neighbors)
\item \code{nu_nhblist} an organized hierarchical list of neurons sorted by their lattice distance from each other. 
\code{nu_nhblist[[i]][[j]]} contains a vector of neuron IDs that are distance j from neuron i on the lattice. 
\item \code{nu_verts} cube whose slices contain the vertices of each lattice tile. 
Slice i contains the vertices of lattice tile i in its rows, in vertex order.  
}
}
