% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somobj_doc.R
\name{set_lattice_fences}
\alias{set_lattice_fences}
\title{Compute and store the U-Matrix fences}
\usage{
SOMobj$set_lattice_fences()
}
\value{
None, data frame \code{fences} is set internally
}
\description{
The U-Matrix fences, when visualized on the SOM lattice, are an easy way to 
determine the similarity between neighboring lattice-neighboring prototypes.  This function 
computes a data frame containing the information required to visiualize such fences, which 
can be accomplished via \link{vis_som_fences}.  

It is usually not necessary to directly call this function as it is invoked inside \link{recall_SOM}.
}
\references{
\insertRef{UMatrix}{SOMDisco}
}
