% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somobj_doc.R
\name{set_nBMU}
\alias{set_nBMU}
\title{Set the number of BMUs}
\usage{
SOMobj$set_nBMU(nBMU)
}
\arguments{
\item{nBMU}{the number of BMUs requested to compute}
}
\value{
None, \code{nBMU} is set internally.
}
\description{
The Best Matching Unit of a sample vector \code{x} is the prototype \eqn{w_j} (and its associated neuron) that 
minimizes Euclidean distance \eqn{d(x,w_j)}.  The SOM requires calculating at least the first BMU during training and recall. 
Calculation of the \code{CADJ} matrix requires calculating and storing the second BMU as well (the second Best Matching Unit).  
This function sets the internal field \code{nBMU}, which specifies the number of BMUs stored in the field \code{BMU} during recall. 
Any number > 2 could be useful for further analysis.  Be default, \code{nBMU} is set = 2 during \link{initialize_SOM}.
}
