% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somobj_doc.R
\name{set_netrng}
\alias{set_netrng}
\title{Set the external/internal network range}
\usage{
SOMobj$set_netrng(ext_min, ext_max, int_min, int_max).
}
\arguments{
\item{ext_min}{min of the external data range (the training data min)}

\item{ext_max}{max of the external data range (the training data max)}

\item{int_min}{min of the internal data range (the prototype min)}

\item{int_max}{max of the internal data range (the prototype max)}
}
\value{
None, scaling ranges are set internally
}
\description{
Training (i.e., prototype updating) is performed in an "internal" network range. 
When presented to the network for training, each data vector is scaled linearly from the external 
data range to the internal data range. This function sets the min/max of both ranges.
}
\details{
Both \code{ext_min} and \code{ext_max} are stored internally 
(as \code{netrng_ext_min} and \code{netrng_ext_max}, respectively) vectors of length \code{d} (data dimension). 
Thus, either can be input as either a length \code{d} vector (to specify a different range of linearly scaling across dimension),  
or as a single number (which will be recycled across dimension to produce a length \code{d} vector). 
\code{int_min} and \code{int_max} (which are stored internally as \code{netrng_int_min} and \code{netrng_int_max}, respectively) 
should be given as a single number (not a vector, varying internal ranges across dimension is not currently supported).  

\code{set_netrng} is called by default during \link{initialize_SOM} using the observed min/max of the data as the external range, 
and 0/1 as the internal range.  

Scaling can be turned off by setting \code{ext_min = int_min = 0} and \code{ext_max = int_max = 1}.
}
