% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somobj_doc.R
\name{set_p}
\alias{set_p}
\title{User-specified initialization of CSOM win frequencies}
\usage{
SOMobj$set_p(p)
}
\arguments{
\item{p}{user-specified win frequency vector. Must have length = \code{nW}}
}
\value{
None, win frequency vector \code{p} is stored internally
}
\description{
This function allows setting of the internal field \code{p}, which stores the CSOM win frequencies for each prototype. 
It is made available mostly for experimentation.
}
