% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somobj_doc.R
\name{set_p_equal}
\alias{set_p_equal}
\title{Equi-probable initialization of CSOM win frequencies}
\usage{
SOMobj$set_p_equal()
}
\value{
None, win frequency vector \code{p} is stored interally.
}
\description{
The Conscience-SOM algorithm of DeSieno attempts a Maximum-Entropy SOM mapping 
by updating a record of each prototype's win frequency. This roughly mimics the proportion of times during 
training that each prototype is selected as some datum's BMU.  
This function initializes the win frequencies for all prototypes to 1 / \code{nW}, which are stored internally in field \code{p}.

Note: this function is called automatically by \link{initialize_SOM}.
}
\references{
\insertRef{DeSieno1988}{SOMDisco}
}
