% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somobj_doc.R
\name{set_W}
\alias{set_W}
\title{Setter function for the prototype matrix \code{W}}
\usage{
SOMobj$set_W(W)
}
\arguments{
\item{W}{matrix of prototype values, with rows following neuron ordering. 
\code{nrow(W)} must equal \code{nW} and \code{ncol(W)} must equal \code{d}, otherwise an error will be returned.}
}
\value{
None, prototype matrix \code{W} is set internally.
}
\description{
Random prototype initialization is performed via \link{set_W_runif}. 
This function allows setting \code{W} to user-prescribed values. 
If the input values reside outside of the internal network range a warning will be issued
}
