% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somobj_doc.R
\name{set_W_runif}
\alias{set_W_runif}
\title{Random uniform initialization of SOM prototypes}
\usage{
SOMobj$set_W_runif()
}
\value{
None, prototype matrix \code{W} is initialized internally.
}
\description{
SOM training typically begins with randomly initialized prototype vectors. 
This function performs this random initialization in the range \code{mid} +/- 5% of the internal network range, 
where \code{mid} is the midpoint of \code{netrng_int_min} and \code{netrng_int_max}. 

To force random sampling according to a specified seed, call R's \link[base]{set.seed} function just prior to calling \code{set_W_runif}

Note: this function is called automatically by \link{initialize_SOM}.
}
