% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somobj_doc.R
\name{set_X_label}
\alias{set_X_label}
\title{Set the training data labels}
\usage{
SOMobj$set_X_label(X_label, ctab)
}
\arguments{
\item{X_label}{a vector (length = \code{nW}) containing character labels for each training observation.}

\item{ctab}{a color table defining the mapping between the unique labels found in \code{X_label} and distinct colors. 
Must be a data frame with columns 'label' and 'color' (where 'color' is in HEX format).}
}
\value{
None, field \code{X_label} is set internally.
}
\description{
If the training data possess labels, this method stores them in the \code{X_label} field, which is used 
to propagate the \code{RF_label} and \code{RF_label_dist} fields during \code{recall_SOM}.
}
\details{
During \code{initialize_SOM} the data labels are set to a default value of "?". This method allows changing these defaults. 
The function \code{build_ctab} can be used to construct a default color table, if needed.
}
