% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somobj_doc.R
\name{tile_interior_point}
\alias{tile_interior_point}
\title{Get a list of interior points in each lattice tile}
\usage{
SOMobj$tile_interior_point(theta, rprop)
}
\arguments{
\item{theta}{the angle at which the (x,y) coordinates should be placed, relative to the tile center}

\item{rprop}{the proportion of the tile radius along angle \code{theta} at which the (x,y) coordinates will be placed. 
Should be < 1 to produce a point inside the lattice tile.}
}
\value{
a matrix (nrow = nW, ncol = 2) containing the (x,y) coordinates of interior points for each tile. 
The rows are in neuron-order.
}
\description{
Handy for visualization, this function returns the (x,y) coordinates of a point inside each 
lattice tile that is along a ray of angle \code{theta} emanating from the tile center. 
The "radius" of a tile is the distance to the tile boundary in direction \code{theta}. 
The (x,y) coordinates are of distance = \code{rprop} x radius from the tile center.
}
