% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somobj_doc.R
\name{update_learning_rates}
\alias{update_learning_rates}
\title{Update the effective learning rates}
\usage{
SOMobj$update_learning_rates()
}
\value{
None, the above learning rate fields are set internally.
}
\description{
The CSOM learning rates (described in \link{set_LRAS}) should be annealed over time. 
This function queries the LRAS given the current training age of the SOM and updates the effective 
\code{alpha}, \code{beta}, \code{gamma}, \code{sigma}, and \code{eta} (the last via a call to \link{calc_eta}) parameters accordingly.  

It is usually not necessary to directly call this function as it is invoked inside \link{train_SOM}.
}
