% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somobj_doc.R
\name{update_p}
\alias{update_p}
\title{Update the CSOM win frequencies}
\usage{
SOMobj$update_p(winner_idx)
}
\arguments{
\item{winner_idx}{the prototype index of the BMU selected during a CSOM training step}
}
\value{
None
}
\description{
The win frequency field \code{p} is updated internally according to DeSieno's CSOM algorithm.

It is usually not necessary to directly call this function as it is invoked inside \link{train_SOM}.
}
