% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{vis_ctab}
\alias{vis_ctab}
\title{View a color table}
\usage{
vis_ctab(ctab, label.cex = 0.9, label.font = 2, nrows_in_display = NULL)
}
\arguments{
\item{ctab}{the color table dataframe to view. Must have fields  \code{$class}, \code{$R}, \code{$G}, \code{$B}.}

\item{label.cex}{the size of labels, as plotted on the visual color grid}

\item{label.font}{the font size of the label, 1=regular, 2 = bold}

\item{nrows_in_display}{sets the number of rows in the visualized color table. Default = NULL invokes automatic generation of this value,
with an attempt to make the resulting color grid as square as possible.}
}
\value{
nothing, only used for visualization
}
\description{
This is a helper function to provide a quick visualization of the labels and their associated colors that are defined
in an input color table.
}
