% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_som.R
\name{vis_som_annotate}
\alias{vis_som_annotate}
\title{Annotate each tile of the SOM lattice}
\usage{
vis_som_annotate(
  SOM,
  add = FALSE,
  text,
  text.cex = 1,
  text.col = "auto",
  text.font = 1,
  text.lightness = 0.4,
  theta = 135,
  rprop = 0,
  active = T,
  subset = NULL,
  change.par = TRUE
)
}
\arguments{
\item{SOM}{a SOM object}

\item{add}{whether to create a new SOM visualization panel (=FALSE, default), or add to an existing one (=TRUE)}

\item{text}{vector of strings used to label each lattice tile}

\item{text.cex}{the text size used for plotting the labels}

\item{text.col}{the color used for plotting the labels. Can be one of: 
\itemize{
\item a single color name, which will be recycled across all lattice tiles
\item a vector of color names to be applied to each tile
\item the reserved keyword "auto", which will set a text color of either white or black, 
depending on the Lightness (from the HSL colorspace) values of any existing tile fill colors. 
This is useful if, for example, some tiles are dark (where black text would not show up well) and others are light (where light text would not show up well). 
}}

\item{text.font}{the font used for plotting the labels}

\item{text.lightness}{the Lightness threshold used to determine whether the text.col in each tile is white or black. 
Only valid if text.col = "auto", default = 0.4. Higher thresholds create more white labels.}

\item{theta}{the angle (in degrees) which specifies the direction of offset of each label, 
relative to each tile center.}

\item{rprop}{the length of the offset of each label, relative to each tile center. 
This should be given as a proportion of the total distance from the center to the boundary of each tile, 
in the direction theta (i.e., \code{rprop} <= 1)}

\item{active}{Optional, if the SOM object has been recalled,
restricts plotting only to active neurons (those whose RF_size > 0).
Default = TRUE.}

\item{subset}{Optional, a vector of neuron indices to restrict the plotting to. 
Default = NULL imposes no restriction (plots whole lattice)}

\item{change.par}{whether to allow the vis function to optimally change the \code{par} of the plot. Default = TRUE. 
If \code{par} is allowed to changed internally, it is always reset upon function exit.}
}
\description{
Annotate each tile of the SOM lattice
}
