% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_som.R
\name{vis_som_fences}
\alias{vis_som_fences}
\title{Visualize SOM fences}
\usage{
vis_som_fences(
  SOM,
  add = FALSE,
  fence.lwd = 2,
  clamp = "none",
  clamp.range = c(0, 1),
  color.mapping = "linear",
  color.palette = c("black", "white"),
  color.range = c(0, 1),
  color.nbins = "none",
  active = T,
  subset = NULL,
  change.par = TRUE
)
}
\arguments{
\item{SOM}{a SOM object}

\item{add}{whether to create a new SOM visualization panel (=FALSE, default), or add to an existing one (=TRUE)}

\item{fence.lwd}{fence line width}

\item{clamp}{controls how values are clamped before unit mapping.  Options are: 
"none" (default), meaning no clamping is used; 
"identity", meaning the clamp.range is used directly; 
"quantile", meaning the clamp.range quantiles are used.}

\item{clamp.range}{sets the clamping range. 
If clamp = 'none', the min/max of values is used. 
If clamp = 'identity', clamp.range should be given as an effective [lo,hi] range to clamp values to. 
If clamp = 'quantile', clamp.range should contain the effective [lo,hi] probabilities whose quantiles define the clamp range. 
Default = c(0,1), which should be changed if clamp = 'identity'.}

\item{color.mapping}{the mapping function used to map values to the range [0,1], which is needed for plotting. 
Default ='linear' for a linear mapping. 
Can also be 'zcdf', which first computes the z-score of the values, then maps the z-score to [0,1] via the Std. Normal CDF.}

\item{color.palette}{the color palette used to represent the fence values. 
This should be given as a character vector of color names which will set the colors which represent the 
the min (first element) and max (last element) of the fence values. Colors for intermediate values between min and max 
will be interpolated via \link[grDevices]{colorRamp}.  
Should have at least two elements, but can have more. In this case, the colors will be interpolated throughout the 
range of those found in \code{color.palette}.  Example:  \code{c('black','yellow','white')} will map the min values to black, 
the mid values to yellow and the max values to white.  
Default = c('black','white').}

\item{color.range}{a range in [0,1] defining what portion of the color scale is visualized. 
This is useful if you want to restrict the plotted color range to the darker or lighter end of the spectrum. 
Default = c(0,1)}

\item{color.nbins}{allows binning of the (unit-mapped) values prior to color assignment. Options are: 
'none' (default), which just maps the values directly to the color range without binning; 
'auto', which computes a histogram of the (unit-mapped) values and uses the resulting bin counts; 
or an integer giving the desired number of bins.}

\item{active}{Optional, if the SOM object has been recalled,
restricts plotting only to active neurons (those whose RF_size > 0).
Default = TRUE.}

\item{subset}{Optional, a vector of neuron indices to restrict the plotting to. 
Default = NULL imposes no restriction (plots whole lattice)}

\item{change.par}{whether to allow the vis function to optimally change the \code{par} of the plot. Default = TRUE. 
If \code{par} is allowed to changed internally, it is always reset upon function exit.}
}
\description{
Visualize SOM fences
}
