% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_som.R
\name{vis_som_label}
\alias{vis_som_label}
\title{Visualize the labels of each RF}
\usage{
vis_som_label(
  SOM,
  add = FALSE,
  text.cex = 1,
  text.font = 1,
  theta = 90,
  rprop = 0.75,
  subset = NULL,
  change.par = TRUE
)
}
\arguments{
\item{SOM}{a SOM object}

\item{add}{whether to create a new SOM visualization panel (=FALSE, default), or add to an existing one (=TRUE)}

\item{subset}{Optional, a vector of neuron indices to restrict the plotting to. 
Default = NULL imposes no restriction (plots whole lattice)}

\item{change.par}{whether to allow the vis function to optimally change the \code{par} of the plot. Default = TRUE. 
If \code{par} is allowed to changed internally, it is always reset upon function exit.}
}
\description{
If the training data is labeled then the SOM prototypes (and their neurons and RFs) will inherit a label 
from the learned mapping. Each prototype is labeled by a plurality vote of the labels of data mapped to their RFs.
}
\details{
This is a wrapped function which calls \code{vis_som_tiles} internally using the internal color table \code{$ctab} 
to map the \code{$RF_labels} to colors. Call \code{$set_ctab} to change this mapping.
}
