% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_som.R
\name{vis_som_labeldist}
\alias{vis_som_labeldist}
\title{Visualize the distribution of labels in each RF}
\usage{
vis_som_labeldist(SOM, add = FALSE, subset = NULL, change.par = TRUE)
}
\arguments{
\item{SOM}{a SOM object}

\item{add}{whether to create a new SOM visualization panel (=FALSE, default), or add to an existing one (=TRUE)}

\item{subset}{Optional, a vector of neuron indices to restrict the plotting to. 
Default = NULL imposes no restriction (plots whole lattice)}

\item{change.par}{whether to allow the vis function to optimally change the \code{par} of the plot. Default = TRUE. 
If \code{par} is allowed to changed internally, it is always reset upon function exit.}
}
\description{
The distribution of data labels mapped to each Receptive Field is visualized as a 
proportional pie chart residing in each lattice tile. This function is helpful for determining the 
organization of SOM neurons: assuming the data labels accurately reflect differences within the dataset, 
prototypes whose receptive fields contain data of the same label are considered 
to be better representative than those whose data member labels are mixed.
}
\details{
The internal color table \code{$ctab} is used to map labels to colors. Call \code{$set_ctab} to change this mapping.
}
