% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_som.R
\name{vis_som_mUMatrix}
\alias{vis_som_mUMatrix}
\title{Visualize a color gradient on SOM lattice tiles}
\usage{
vis_som_mUMatrix(
  SOM,
  add = FALSE,
  grad.clamp = "quantile",
  grad.clamp.range = c(0.05, 0.95),
  grad.color.mapping = "linear",
  grad.color.palette = c("black", SOMDisco::tableau20("lightblue")),
  grad.color.range = c(0, 1),
  grad.color.nbins = "auto",
  fence.lwd = 2,
  fence.clamp = "quantile",
  fence.clamp.range = c(0.05, 0.95),
  fence.color.mapping = "linear",
  fence.color.palette = c("black", "white"),
  fence.color.range = c(0, 1),
  fence.color.nbins = "auto",
  subset = NULL,
  change.par = TRUE
)
}
\arguments{
\item{SOM}{a SOM object}

\item{add}{whether to create a new SOM visualization panel (=FALSE, default), or add to an existing one (=TRUE)}

\item{grad.clamp}{see \code{vis_som_gradient}}

\item{grad.clamp.range}{see \code{vis_som_gradient}}

\item{grad.color.mapping}{see \code{vis_som_gradient}}

\item{grad.color.palette}{see \code{vis_som_gradient}}

\item{grad.color.range}{see \code{vis_som_gradient}}

\item{grad.color.nbins}{see \code{vis_som_gradient}}

\item{fence.lwd}{see \code{vis_som_fences}}

\item{fence.clamp}{see \code{vis_som_fences}}

\item{fence.clamp.range}{see \code{vis_som_fences}}

\item{fence.color.mapping}{see \code{vis_som_fences}}

\item{fence.color.palette}{see \code{vis_som_fences}}

\item{fence.color.range}{see \code{vis_som_fences}}

\item{fence.color.nbins}{see \code{vis_som_fences}}

\item{subset}{Optional, a vector of neuron indices to restrict the plotting to. 
Default = NULL imposes no restriction (plots whole lattice)}

\item{change.par}{whether to allow the vis function to optimally change the \code{par} of the plot. Default = TRUE. 
If \code{par} is allowed to changed internally, it is always reset upon function exit.}
}
\description{
Visualize a color gradient on SOM lattice tiles
}
\details{
This is a wrapper function which layers \code{vis_som_fences} (called with the given \code{fence.*} parameters)  atop 
\code{vis_som_gradient} (called with the given \code{grad.*} parameters).
}
\references{
\insertRef{UMatrix}{SOMDisco}
\insertRef{MerenyiJainVillmann}{SOMDisco}
}
