% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_som.R
\name{vis_som_neurons}
\alias{vis_som_neurons}
\title{Visualize SOM neurons on the lattice}
\usage{
vis_som_neurons(
  SOM,
  add = FALSE,
  nu.pch = 16,
  nu.cex = 1,
  nu.col = "black",
  active = TRUE,
  subset = NULL,
  change.par = TRUE
)
}
\arguments{
\item{SOM}{a SOM object}

\item{add}{whether to create a new SOM visualization panel (=FALSE, default), or add to an existing one (=TRUE)}

\item{nu.pch}{neuron marker type, default = 16}

\item{nu.cex}{neuron marker size, default = 1,}

\item{nu.col}{neuron marker color, default = "black"}

\item{active}{Optional, if the SOM object has been recalled,
restricts plotting only to active neurons (those whose RF_size > 0).
Default = TRUE.}

\item{subset}{Optional, a vector of neuron indices to restrict the plotting to. 
Default = NULL imposes no restriction (plots whole lattice)}

\item{change.par}{whether to allow the vis function to optimally change the \code{par} of the plot. Default = TRUE. 
If \code{par} is allowed to changed internally, it is always reset upon function exit.}
}
\description{
Visualize SOM neurons on the lattice
}
\details{
See ?points or ?text for help with plotting parameters cex, col, font, etc
}
