% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_som.R
\name{vis_som_prototypes}
\alias{vis_som_prototypes}
\title{Visualize SOM prototypes at their lattice locations}
\usage{
vis_som_prototypes(
  SOM,
  add = FALSE,
  netrng = "int",
  gutter = 0.05,
  wgt.lwd = 1,
  wgt.col = tableau20("blue"),
  axes = T,
  axes.lwd = 0.5,
  axes.col = "black",
  axes.nticksx = 4,
  axes.nticksy = 4,
  active = T,
  subset = NULL,
  change.par = TRUE
)
}
\arguments{
\item{SOM}{a SOM object}

\item{add}{whether to create a new SOM visualization panel (=FALSE, default), or add to an existing one (=TRUE)}

\item{netrng}{whether to plot the prototypes in internal ("int") or external ("ext") network range. Default = "int".}

\item{gutter}{the proportion of the tile width / height used for plotting axes in each tile. 
The actual space occupied by the axes = gutter * min(tile_width, tile_height)}

\item{wgt.lwd}{the line width used for plotting the weight vectors}

\item{wgt.col}{the color used for plotting the weight vectors}

\item{axes}{boolean, whether to plot individual axes inside each lattice tile}

\item{axes.lwd}{the line width used for plotting axes}

\item{axes.col}{the color used for plotting axes}

\item{axes.nticksx}{the number of ticks on x-axis inside each lattice tile}

\item{axes.nticksy}{the number of ticks on y-axis inside each lattice tile}

\item{active}{Optional, if the SOM object has been recalled,
restricts plotting only to active neurons (those whose RF_size > 0).
Default = TRUE.}

\item{subset}{Optional, a vector of neuron indices to restrict the plotting to. 
Default = NULL imposes no restriction (plots whole lattice)}

\item{change.par}{whether to allow the vis function to optimally change the \code{par} of the plot. Default = TRUE. 
If \code{par} is allowed to changed internally, it is always reset upon function exit.}
}
\description{
Visualize SOM prototypes at their lattice locations
}
