% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_som.R
\name{vis_som_tiles}
\alias{vis_som_tiles}
\title{Visualize SOM lattice tiles}
\usage{
vis_som_tiles(
  SOM,
  add = FALSE,
  border.lwd = 1,
  border.lty = 1,
  border.col = "black",
  fill = NULL,
  active = TRUE,
  subset = NULL,
  change.par = TRUE
)
}
\arguments{
\item{SOM}{a SOM object}

\item{add}{whether to create a new SOM visualization panel (=FALSE, default), or add to an existing one (=TRUE)}

\item{border.lwd}{tile border line width, default = 1. Set = 0 to suppress plotting borders.}

\item{border.lty}{tile border line type, default = 1}

\item{border.col}{tile border color, default = "black". 
If border.col = 'fill' and input fill is given, borders will inherit fill colors.}

\item{fill}{Optional, controls fill color of plotted tiles. 
If given as a single color name (e.g., "black"), all tiles will be colored with this color. 
If given as a vector of color names with length = SOM$nW, each tile will be colored separately. 
Default = NULL means no fill}

\item{active}{Optional, if the SOM object has been recalled,
restricts plotting only to active neurons (those whose RF_size > 0).
Default = TRUE.}

\item{subset}{Optional, a vector of neuron indices to restrict the plotting to. 
Default = NULL imposes no restriction (plots whole lattice)}

\item{change.par}{whether to allow the vis function to optimally change the \code{par} of the plot. Default = TRUE. 
If \code{par} is allowed to changed internally, it is always reset upon function exit.}
}
\description{
Visualize SOM lattice tiles
}
