% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_som.R
\name{vis_som_TopoView}
\alias{vis_som_TopoView}
\title{TopoView Visualization}
\usage{
vis_som_TopoView(
  SOM,
  ADJ,
  add = F,
  nu.pch = 16,
  nu.cex = 1,
  nu.col = "black",
  edge.col = "darkorange",
  edge.lwd_range = c(1, 5),
  active = T,
  subset = NULL
)
}
\arguments{
\item{SOM}{a SOM object}

\item{ADJ}{an adjacency matrix defining edges connecting SOM neurons (or prototypes)}

\item{add}{whether to create a new plotting device (=FALSE, default), or add to an existing one (=TRUE)}

\item{nu.pch}{see \code{vis_som_neurons}}

\item{nu.cex}{see \code{vis_som_neurons}}

\item{nu.col}{see \code{vis_som_neurons}}

\item{edge.lwd_range}{the min/max range of the plotted CONN edges, Default = c(1, 5).}

\item{active}{Optional, if the SOM object has been recalled,
restricts plotting (of vertices and edges) only to active neurons (those whose RF_size > 0).
Default = TRUE.}

\item{subset}{Optional, a vector of neuron indices to restrict the plotting to. 
Default = NULL imposes no restriction (plots whole lattice)}

\item{edge.color}{line color of plotted edges, default = "darkorange".}
}
\description{
TopoView Visualization
}
\details{
The TopoView visualization is documented in \code{TopoRNet::vis_TopoView}.  If the input \code{ADJ} is weighted and 
\code{edge.lwd_range} spans a non-empty set, the visualized edge widths will represent the edge weights (larger weights = thicker edges).
}
\references{
\insertRef{Merenyietal2009}{SOMDisco}
}
