% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_som.R
\name{vis_som_training}
\alias{vis_som_training}
\title{Visualize monitoring measure of SOM training}
\usage{
vis_som_training(SOM, vis.SOM = T, vis.mtr = T)
}
\arguments{
\item{SOM}{a SOM object}

\item{vis.SOM}{boolean, whether to show the SOM with RF_size gradient, default = T}

\item{vis.mtr}{boolean, whether to show the monitoring history, if available. Default = T}
}
\description{
Visualize monitoring measure of SOM training
}
